+++ native/plugins/foobar2000/resource.h
@@ -1,0 +1,10 @@
+// Copyright 2026 Lanes Audio
+// SPDX-License-Identifier: LicenseRef-Apache-2.0-with-Commons-Clause-1.0
+#pragma once
+
+#define IDD_OPTILAB_FOOBAR2000_CONFIG 201
+#define IDC_OPTILAB_MODE 2001
+#define IDC_OPTILAB_INPUT_DRIVE 2002
+#define IDC_OPTILAB_AUTO_ADAPT 2003
+#define IDC_OPTILAB_RESTORE_DEFAULTS 2004
+#define IDC_OPTILAB_AUTO_ADAPT_VALUE 2005